refactor(#60): transaction class and main method refactored
